make a libs directory for the small things
